+++ CHANGES.rst
@@ -232,6 +232,10 @@ astropy.convolution
 astropy.coordinates
 ^^^^^^^^^^^^^^^^^^^
 
+- Deprecated ``recommended_units`` for representations. These were used to
+  ensure that any angle was presented in degrees in sky coordinates and
+  frames. This is more logically done in the frame itself. [#6858]
+
 astropy.cosmology
 ^^^^^^^^^^^^^^^^^